🔨 Add pdm_build.py compatible with both typer and typer-cli from the same pyproject.toml
